# public/apiDocsV1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo",
    "description": "",
    "contact": {
      "email": ""
    },
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "User",
      "description": "Operations about user",
      "externalDocs": {
        "description": "",
        "url": ""
      }
    },
    {
      "name": "Admin",
      "description": "Operations about Admin",
      "externalDocs": {
        "description": "",
        "url": ""
      }
    },
    {
      "name": "Widget",
      "description": "Operations about Widget",
      "externalDocs": {
        "description": "",
        "url": ""
      }
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/create": {
      "post": {
        "deprecated": true,
        "tags": [
          "User"
        ],
        "description": "Signup",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Email",
            "required": true,
            "type": "string"
          },
         
          {
            "name": "password",
            "in": "formData",
            "description": "Password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    
    "/user/forgotPassword": {
      "post": {
        "deprecated": true,
        "tags": [
          "User"
        ],
        "description": "forgot Password",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
  
   
    "/user/editProfile": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "editProfile",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "",
            "required": false,
            "type": "string"
          }
       
         ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    },
    "/user/deleteProfile": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "deleteProfile",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": [
              "Username: ",
              "Password: "
            ]
          }
        ]
      }
    }
  
   
   
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  }
}